Clean codes with cookie_services.py and user_services.py

+++ cloud/api/src/services/call_users.py
@@ -1,4 +1,5 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, status, HTTPException, Request
+from models.user_model import User
 from config.db_config import ConfigDB
 
 
@@ -28,28 +29,46 @@ def get_users() -> list[dict[str, str | list[str] | int]]:
 
 
 @router.delete("/del_users/{idUsers}")
-def del_users(idUsers: str) -> None:
-    # Todo : On doit impérativement vérifier le rôle de l'utilisateur avec le cookie. Seuls les admins ont le droit.
+def del_users(request: Request, idUsers: str) -> None:
     DB = ConfigDB()
     cursor = DB.get_db_cursor()
-    SQL_query = (
-        f"DELETE FROM USERS WHERE idusers='{idUsers}'"
-    )
-    cursor.execute(SQL_query)
-    DB.connector.commit()
-    cursor.close()
-    DB.connector.close()
+    user_role = User().get_current_user_role(request.cookies.get("ICARUS-Login"), cursor)
+    if user_role == 1:
+        SQL_query = (
+            f"DELETE FROM USERS WHERE idusers='{idUsers}'"
+        )
+        cursor.execute(SQL_query)
+        DB.connector.commit()
+        cursor.close()
+        DB.connector.close()
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid profile.",
+        )
+
 
 @router.post("/update_users/")
-def update_users(idUsers: str, name: str, forename: str, email: str, role: str) -> None:
-    # TODO : Attention aux cookies!
-    role = 1 if role == "Admin" else 2
+def update_users(request: Request, idUsers: str, name: str, forename: str, email: str, role: str) -> None:
     DB = ConfigDB()
     cursor = DB.get_db_cursor()
-    SQL_query = (
-        f"UPDATE USERS SET name = '{name}', forename = '{forename}', email = '{email}', role = '{role}' WHERE idusers = '{idUsers}'"
-    )
-    cursor.execute(SQL_query)
-    DB.connector.commit()
-    cursor.close()
-    DB.connector.close()
+    user_role = User().get_current_user_role(request.cookies.get("ICARUS-Login"), cursor)
+    if user_role == 1 or request.cookies.get("ICARUS-Login") == User().get_user(cursor, id_users=idUsers).cookie:
+        # If the request come from a User, he can't set the user_role to Admin.
+        # Technically, he can't access the webpage to this, but he can still tinker the request himself.
+        if user_role == 2:
+            role = 2
+        else:
+            role = 1 if role == "Admin" else 2
+        SQL_query = (
+            f"UPDATE USERS SET name = '{name}', forename = '{forename}', email = '{email}', role = '{role}' WHERE idusers = '{idUsers}'"
+        )
+        cursor.execute(SQL_query)
+        DB.connector.commit()
+        cursor.close()
+        DB.connector.close()
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid username or password",
+        )
